Factored terminal state into its own header.
